had to recreate game to fix it, so i also got rid of all the old files, and made new ones
