--- Assets/ClockIn-MG/Card.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bb03d5cfbdd37af4a9d092080c80c336
