--- Assets/ClockIn-MG/CardReader.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 474591a25fba67d49a74475304e5a194
