--- Assets/ClockIn-MG/ClockIn.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: af8695e7154d29d43b1b9bdf57e4ae76
