--- Assets/ClockIn-MG/ClockInEventManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 130b87704b326d847a752679dac11248
