--- Assets/ClockIn-MG/EventManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2e4dd45adc84a9f419e35432f3be90da
