--- Assets/ClockIn-MG/Movement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 40146c9286548b74ab900ad7a71066f3
