+++ Assets/ClockIn-MGPrime.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53630c9c62c6c894a885c2e705e30560
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
