+++ Assets/ClockIn-MGPrime/Card.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7966436c853dade439daa4d2bd7c86c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
