+++ Assets/ClockIn-MGPrime/CardClockInEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dde76deee7f273f48a88fe67f733da07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
