+++ Assets/ClockIn-MGPrime/CardMovement.cs
@@ -2,9 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Movement : MonoBehaviour
+public class CardMovement : MonoBehaviour
 {
-
     private float yPos;
     // Start is called before the first frame update
     void Start()
