+++ Assets/ClockIn-MGPrime/CardMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 686ba090d632daa4aa2336a3ae10bc6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
