+++ Assets/ClockIn-MGPrime/CardReader.cs
@@ -5,21 +5,12 @@ using UnityEngine;
 public class CardReader : MonoBehaviour
 {
     // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
 
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
     void OnTriggerEnter(Collider collisionObject)
     {
         if (collisionObject.name == "Card")
         {
-            EventManager.RunCardSwipe();
+            EventManagerMain.RunCardSwipe();
         }
     }
 }
