+++ Assets/ClockIn-MGPrime/CardReader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67592d0f1a0b7604f9af55751186772c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
