+++ Assets/ClockIn-MGPrime/CardReaderColor.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 327763d015b283e4eb4715940bf16764
-DefaultImporter:
+guid: 8d68d833f6b938943bc4e6aac81a6ed2
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
